FEAT: block model
