+++ src/model/IBlock.java
@@ -1,0 +1,12 @@
+package model;
+
+import java.awt.Color;
+
+class IBlock extends Block {
+    public IBlock() {
+        block = new int[][] {
+                { 1, 1, 1, 1 },
+        };
+        color = Color.RED;
+    }
+}
