+++ src/model/LBlock.java
@@ -1,0 +1,14 @@
+package model;
+
+import java.awt.Color;
+
+class LBlock extends Block {
+    public LBlock() {
+        block = new int[][] {
+                { 1, 0 },
+                { 1, 0 },
+                { 1, 1 },
+        };
+        color = Color.BLUE;
+    }
+}
