+++ src/model/OBlock.java
@@ -1,0 +1,13 @@
+package model;
+
+import java.awt.Color;
+
+class OBlock extends Block {
+    public OBlock() {
+        block = new int[][] {
+                { 1, 1 },
+                { 1, 1 },
+        };
+        color = Color.YELLOW;
+    }
+}
