+++ src/model/TBlock.java
@@ -1,0 +1,13 @@
+package model;
+
+import java.awt.Color;
+
+class TBlock extends Block {
+    public TBlock() {
+        block = new int[][] {
+                { 0, 1, 0 },
+                { 1, 1, 1 },
+        };
+        color = Color.CYAN;
+    }
+}
